feat: Moved useAuth to hooks directory
