+++ src/routes/sign_up/-sign_up.tsx
@@ -2,13 +2,22 @@ import { FormEvent, useState } from "react"
 import { graphql } from "@src/__generated__/gql"
 import { useMutation } from "@apollo/client"
 import { Link, useRouter } from "@tanstack/react-router"
-import { ME } from "@src/useAuth"
+import { ME } from "@src/hooks/useAuth"
 import {
   CreateUserMutation,
   CreateUserMutationVariables,
   MeQuery,
 } from "@src/__generated__/graphql"
-import { Container, Title, Text, Anchor, Paper, TextInput, PasswordInput, Button } from "@mantine/core"
+import {
+  Container,
+  Title,
+  Text,
+  Anchor,
+  Paper,
+  TextInput,
+  PasswordInput,
+  Button,
+} from "@mantine/core"
 
 const CREATE_USER = graphql(`
   mutation CreateUser($input: UserInput) {
@@ -56,47 +65,46 @@ export const SignupForm = () => {
 
   return (
     <Container size={420} my={40}>
-    <Title ta="center">
-      Create an account
-    </Title>
-    <Text c="dimmed" size="sm" ta="center" mt={5}>
-      Already have an account?{' '}
-      <Link to="/login">
-        <Anchor size="sm" component="button">
-          Login
-        </Anchor>
-      </Link>
-    </Text>
+      <Title ta="center">Create an account</Title>
+      <Text c="dimmed" size="sm" ta="center" mt={5}>
+        Already have an account?{" "}
+        <Link to="/login">
+          <Anchor size="sm" component="button">
+            Login
+          </Anchor>
+        </Link>
+      </Text>
 
-    <Paper withBorder shadow="md" p={30} mt={30} radius="md">
-      <form onSubmit={onSubmit}>
-        <TextInput
-          label="Username"
-          placeholder="john_smith"
-          required onChange={e => setUsername(e.target.value)}
-          value={username}
-        />
-        <PasswordInput
-          label="Password"
-          placeholder="Your password"
-          required
-          mt="md"
-          onChange={e => setPassword(e.target.value)}
-          value={password}
-        />
-        <PasswordInput
-          label="Confirm Password"
-          placeholder="confirm password"
-          required
-          mt="md"
-          onChange={e => setConfirmPassword(e.target.value)}
-          value={confirmPassword}
-        />
-        <Button type="submit" fullWidth mt="xl">
-          Sign in
-        </Button>
-      </form>
-    </Paper>
-  </Container>
+      <Paper withBorder shadow="md" p={30} mt={30} radius="md">
+        <form onSubmit={onSubmit}>
+          <TextInput
+            label="Username"
+            placeholder="john_smith"
+            required
+            onChange={(e) => setUsername(e.target.value)}
+            value={username}
+          />
+          <PasswordInput
+            label="Password"
+            placeholder="Your password"
+            required
+            mt="md"
+            onChange={(e) => setPassword(e.target.value)}
+            value={password}
+          />
+          <PasswordInput
+            label="Confirm Password"
+            placeholder="confirm password"
+            required
+            mt="md"
+            onChange={(e) => setConfirmPassword(e.target.value)}
+            value={confirmPassword}
+          />
+          <Button type="submit" fullWidth mt="xl">
+            Sign in
+          </Button>
+        </form>
+      </Paper>
+    </Container>
   )
 }
